+++ packages/orq-rc/RELEASES.md
@@ -549,3 +549,13 @@ Based on:
 - [python v3.2.0-rc.17] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.17] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.17 - packages/orq-rc
+
+## 2025-02-27 08:03:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.505.1 (2.531.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.18] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.18] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.18 - packages/orq-rc